+++ jirkov.vladimir/P3/main.cpp
@@ -3,8 +3,7 @@
 #include "inputString.hpp"
 #include "checkDuplicatesChars.hpp"
 
-int main()
-{
+int main(){
   size_t scope = 10;
   size_t size = 0;
   char* string = inputString(size, scope);
@@ -13,7 +12,7 @@ int main()
     std::cerr << "String reading error\n";
     return 1;
   }
-  size_t Result = checkDuplicatesChars(string, size);
+  int Result = checkDuplicatesChars(string, size);
   std::cout << Result << '\n';
   delete[] string;
   return 0;
